refactor: Remove ragAPI.js and consolidate utility functions into retrievalAPI.js
