+++ src/app/api/chatAPI.js
@@ -1,5 +1,4 @@
 // Configuration API 호출 함수들을 관리하는 파일
-import { devLog } from '@/app/utils/logger';
 import { API_BASE_URL } from '@/app/config.js';
 
 /**
